fix: call the correct ArrayList.remove method while removing the targetId
Here, the `targetId` is cast into Integer to enforce this

+++ library/src/main/java/android/transitions/everywhere/Transition.java
@@ -1070,7 +1070,7 @@ public abstract class Transition implements Cloneable {
      */
     public Transition removeTarget(int targetId) {
         if (targetId > 0) {
-            mTargetIds.remove(targetId);
+            mTargetIds.remove((Integer) targetId);
         }
         return this;
     }
